+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Square bracket attr reader like `person[:name]`
+
 ## [3.4.0] - 2020-08-01
 ### Added
 - Value objects can be instantiated from any object that responds to `#to_h`.